+++ src/ResourceManager/Sql/Commands.Sql/Failover Group/Cmdlet/NewAzureSqlFailoverGroup.cs
@@ -15,7 +15,6 @@
 using Hyak.Common;
 using Microsoft.Azure.Commands.ResourceManager.Common.Tags;
 using Microsoft.Azure.Commands.Sql.FailoverGroup.Model;
-using Microsoft.Azure.Management.Sql.Models;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
@@ -65,12 +64,10 @@ namespace Microsoft.Azure.Commands.Sql.FailoverGroup.Cmdlet
         /// <summary>
         /// Gets or sets the grace period with data loss for the Sql Azure Failover Group.
         /// </summary>
-        [Alias("GracePeriodWithDataLossHours")]
         [Parameter(Mandatory = false,
             HelpMessage = "The window of grace period that we tolerate with data loss during a failover operation for the failover group.")]
-        [Obsolete("This parameter is only for backwards compatibility; User should use 'GracePeriodWithDataLossHours' instead.")]
         [ValidateNotNullOrEmpty]
-        public int GracePeriodWithDataLossHour { get; set; }
+        public int GracePeriodWithDataLossHours { get; set; }
 
         /// <summary>
         /// Gets or sets the failover policy for read only endpoint of theSql Azure Failover Group.
@@ -79,14 +76,6 @@ namespace Microsoft.Azure.Commands.Sql.FailoverGroup.Cmdlet
             HelpMessage = "The failover policy for read only endpoint of the failover group.")]
         [ValidateNotNullOrEmpty]
         public AllowReadOnlyFailoverToPrimary AllowReadOnlyFailoverToPrimary { get; set; }
-
-        /// <summary> 
-        /// Gets or sets the tags associated with the Azure SQL Database Failover Group 
-        /// </summary> 
-        [Parameter(Mandatory = false,
-            HelpMessage = "The tags to associate with the Azure SQL Database Failover Group")]
-        [Obsolete("This parameter is only for backwards compatibility; User should not need to pass in this parameter.")]
-        public Hashtable Tag { get; set; }
 
         /// <summary>
         /// Get the entities from the service
@@ -126,8 +115,7 @@ namespace Microsoft.Azure.Commands.Sql.FailoverGroup.Cmdlet
         {
             string location = ModelAdapter.GetServerLocation(ResourceGroupName, ServerName);
             List<AzureSqlFailoverGroupModel> newEntity = new List<AzureSqlFailoverGroupModel>();
-
-            newEntity.Add(new AzureSqlFailoverGroupInputModel()
+            newEntity.Add(new AzureSqlFailoverGroupModel()
             {
                 ResourceGroupName = ResourceGroupName,
                 ServerName = ServerName,
@@ -135,13 +123,10 @@ namespace Microsoft.Azure.Commands.Sql.FailoverGroup.Cmdlet
                 FailoverGroupName = FailoverGroupName,
                 PartnerResourceGroupName = MyInvocation.BoundParameters.ContainsKey("PartnerResourceGroupName") ? PartnerResourceGroupName : ResourceGroupName,
                 PartnerServerName = PartnerServerName,
-                ReadWriteFailoverPolicy = FailoverPolicy.ToString(),
-#pragma warning disable 0618
-                FailoverWithDataLossGracePeriodHours = GracePeriodWithDataLossHour,
-                ReadOnlyFailoverPolicy = AllowReadOnlyFailoverToPrimary.ToString()
-#pragma warning restore 0618
+                ReadWriteFailoverPolicy  = MyInvocation.BoundParameters.ContainsKey("FailoverPolicy") ? FailoverPolicy.ToString() : FailoverPolicy.Manual.ToString(),
+                FailoverWithDataLossGracePeriodHours = GracePeriodWithDataLossHours,
+                ReadOnlyFailoverPolicy = MyInvocation.BoundParameters.ContainsKey("AllowReadOnlyFailoverToPrimary") ? AllowReadOnlyFailoverToPrimary.ToString() : AllowReadOnlyFailoverToPrimary.Disabled.ToString(),
             });
-
             return newEntity;
         }
 
@@ -153,7 +138,7 @@ namespace Microsoft.Azure.Commands.Sql.FailoverGroup.Cmdlet
         protected override IEnumerable<AzureSqlFailoverGroupModel> PersistChanges(IEnumerable<AzureSqlFailoverGroupModel> entity)
         {
             return new List<AzureSqlFailoverGroupModel>() {
-                ModelAdapter.UpsertFailoverGroup(entity.First() as AzureSqlFailoverGroupInputModel)
+                ModelAdapter.UpsertFailoverGroup(entity.First())
             };
         }
     }